update ‘create set’ screen.

+++ Memento/Memento/SetTableViewCell.m
@@ -20,7 +20,7 @@
 
 -(void)configureWithTitle:(NSString *)title termsCount:(NSUInteger)count author:(NSString *)author {
     self.titleLabel.text = title;
-    self.countTermsLabel.text = [NSString stringWithFormat:@"%lu terms", count];
+    self.countTermsLabel.text = [NSString stringWithFormat:@"%lu terms", (unsigned long)count];
     self.authorLabel.text = author;
 }
